Improved code robustness
(Only applied in Source/Engine/DX12. Other changes are to reflect code compilation. Still need to work on other folders)

* Changed class definition convention. Parameters used to be on top, and then functions. Now it follows this convention:

class
{
	friends
	constructors/destructors
	member functions
	static functions
	member variables
}

All members should be grouped based on their member type first (as shown above), then grouped by their access specifier in this order: private, protected, public
Maybe would also make sense to separate based on virtual, pure virtual, non-virtual.

* Made some classes' destructors private when it makes sense, and use friends class instead.
* Removed default empty comstructor/destructors. Replaced it with either using "Parent::Parent" or "= default"
* Inlined Getters/Setters where it makes sense
* Fixed some warnings

* Found a nasty bug while doing this (this is why robustness is worth the effort):

I was incorrectly calling a virtual function inside a destructor (See previous ~DX12Resource()).
This is actually an undefinied behaviour and worked by pure chance.
I refactored eveything, since it was a design mistake to begin with.
Since we called the Init() function to create resources, it doesn't make any sense to let the destructor release those resources. Instead we should epxlicitely call the Release() function.
The robustness part I did here is that instead of overriding Release() and calling the parent's Release() function (which could lead to leaks if we forget that), there is a new OnRelease() function that can be overridden,  and is called by the base class in the Release() function. Robustness!

+++ Source/Engine/Utils/Exceptions.cpp
@@ -19,7 +19,7 @@ bool HandleAssert(const char* inFileName, int inLineNumber, const char* inExpres
 	Trace("******************************************************************************************************************************************************");
 	Trace("*                                                            Assert failed                                                                           *");
 	Trace("******************************************************************************************************************************************************");
-	int expression_size = 142 - std::strlen(inFileName) - ::log10((float) inLineNumber);
+	size_t expression_size = 142 - std::strlen(inFileName) - static_cast<size_t>(::log10(static_cast<float>(inLineNumber)));
 	Trace("* %s(%d): %-*s *", inFileName, inLineNumber, expression_size, inExpression);
 	if (inMessage != nullptr)
 		Trace("* %-*s *", 146, inMessage);
